Moved Map.js and map-components to pages/comparison
