+++ src/components/pages/comparison/map-components/DataDisplay.scss
