+++ src/components/pages/comparison/map-components/FavCard.js
