+++ src/components/pages/comparison/map-components/FavoriteButton.js
