+++ src/components/pages/comparison/map-components/Map.scss
@@ -1236,7 +1236,7 @@
             }
         }
         .search-bar{
-            background-image:url('../dashboard/assets/search.svg');
+            background-image:url('../../../dashboard/assets/search.svg');
             background-position:6px 5.5px;
             background-repeat: no-repeat;
             padding:16px 33px;
