+++ src/components/pages/comparison/map-components/MapFooter.js
