+++ src/components/pages/comparison/map-components/MapSearch.js
