+++ src/components/pages/comparison/map-components/Markers.js
