+++ src/components/pages/comparison/map-components/data-categories/Housing.js
