+++ src/components/pages/comparison/map-components/data-categories/Jobs.js
