+++ src/components/pages/comparison/map-components/data.js
