+++ src/components/pages/comparison/map-components/icons/activepin.png
