+++ src/components/pages/comparison/map-components/icons/pin.png
